// apps/builder/src/api/ws/interface.ts
import { ComponentTreeNode } from "@flowagent-public/public-types"
import { TextSignal, TextTarget } from "@/api/ws/textSignal"

export type RoomType = "dashboard" | "app" | "ai-agent"

export interface Room {
  wsURL: string
}

export interface Broadcast {
  type: string
  payload: any
}

export interface Callback<T> {
  broadcast: Broadcast
  // extra data
  data: T
  // string
  errorMessage: string
  // 0 success, not zero error
  errorCode: number
  target: TextTarget
  signal: TextSignal
}

export interface FLOWAGENTWebSocketComponentPayload {
  before: {
    displayName: string
  }
  after: ComponentTreeNode
}

export enum FLOWAGENT_WEBSOCKET_STATUS {
  INIT = "INIT",
  CONNECTING = "CONNECTING",
  CONNECTED = "CONNECTED",
  CLOSED = "CLOSED",
  FAILED = "FAILED",
  LOCKING = "LOCKING",
}

export enum FLOWAGENT_WEBSOCKET_CONTEXT {
  DASHBOARD = "DASHBOARD",
  APP = "APP",
  APP_BINARY = "APP_BINARY",
  AI_AGENT = "AI_AGENT",
}